research_agent/requirements.txt
langchain
langchain-anthropic
langchain-google-community
google-generativeai
openai
faiss-cpu
python-dotenv
